FEI-5116: Refactor tsconfig files to properly spearate typecheck and build configs
