+++ packages/wonder-stuff-render-environment-jsdom/tsconfig-build.json
@@ -1,0 +1,15 @@
+{
+  "exclude": [
+    "dist", "examples"
+  ],
+  "extends": "../tsconfig-shared.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src",
+  },
+  "references": [
+    {"path": "../wonder-stuff-core/tsconfig-build.json"},
+    {"path": "../wonder-stuff-server/tsconfig-build.json"},
+    {"path": "../wonder-stuff-render-server/tsconfig-build.json"},
+  ]
+}
