+++ packages/wonder-stuff-render-server/src/requests-from-cache.ts
@@ -57,6 +57,26 @@ export const asUncachedRequest = (
     return superagentRequest as any;
 };
 
+// TODO(kevinb): Figure out how to move this into its own .d.ts file and have it
+// work with both `yarn typecheck` and `yarn build:types`.
+declare module "superagent" {
+    interface SuperAgentRequest {
+        cacheWhenEmpty(cacheWhenEmpty: boolean): this;
+        doQuery(doQuery: boolean): this;
+        expiration(expiration?: number): this;
+        forceUpdate(forceUpdate?: boolean): this;
+        pruneQuery(pruneQuery: Array<string>): this;
+        pruneHeader(pruneHeader: Array<string>): this;
+        prune(
+            prune: (
+                response: Response,
+                gutResponse: (response: Response) => any,
+            ) => any,
+        ): this;
+        responseProp(responseProp: string): this;
+    }
+}
+
 /**
  * Turn unbuffered, uncached request into cached request with or without buffer.
  *
