--- packages/wonder-stuff-render-server/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "exclude": [
-    "dist", "examples"
-  ],
-  "extends": "../tsconfig-shared.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-  },
-  "references": [
-    {"path": "../wonder-stuff-core"},
-    {"path": "../wonder-stuff-server"}
-  ]
-}
