--- packages/wonder-stuff-render-server/types/.eslintrc.js
@@ -1,11 +1,0 @@
-/* eslint-disable import/no-commonjs */
-module.exports = {
-    rules: {
-        "import/no-unassigned-import": [
-            "error",
-            {
-                allow: ["superagent"],
-            },
-        ],
-    },
-};
