+++ packages/wonder-stuff-sentry/tsconfig-build.json
@@ -6,5 +6,6 @@
     "rootDir": "src",
   },
   "references": [
+    {"path": "../wonder-stuff-core/tsconfig-build.json"},
   ]
 }
