+++ packages/wonder-stuff-server/tsconfig-build.json
@@ -2,9 +2,10 @@
   "exclude": ["dist"],
   "extends": "../tsconfig-shared.json",
   "compilerOptions": {
-    "outDir": "dist",
+    "outDir": "./dist",
     "rootDir": "src",
   },
   "references": [
+    {"path": "../wonder-stuff-core/tsconfig-build.json"},
   ]
 }
