+++ packages/wonder-stuff-testing/tsconfig-build.json
@@ -2,10 +2,10 @@
   "exclude": ["dist"],
   "extends": "../tsconfig-shared.json",
   "compilerOptions": {
-    "outDir": "./dist",
+    "outDir": "dist",
     "rootDir": "src",
   },
   "references": [
-    {"path": "../wonder-stuff-core"}
+    {"path": "../wonder-stuff-core/tsconfig-build.json"},
   ]
 }
