+++ tsconfig-build.json
@@ -5,13 +5,13 @@
   // in each package directory.
   "files": [],
   "references": [
-    {"path": "./packages/wonder-stuff-core"},
-    {"path": "./packages/wonder-stuff-i18n"},
-    {"path": "./packages/wonder-stuff-render-environment-jsdom"},
-    {"path": "./packages/wonder-stuff-render-server"},
-    {"path": "./packages/wonder-stuff-sentry"},
-    {"path": "./packages/wonder-stuff-server"},
-    {"path": "./packages/wonder-stuff-testing"},
-    {"path": "./packages/wonder-stuff-ci"},
+    {"path": "./packages/wonder-stuff-core/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-i18n/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-render-environment-jsdom/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-render-server/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-sentry/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-server/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-testing/tsconfig-build.json"},
+    {"path": "./packages/wonder-stuff-ci/tsconfig-build.json"},
   ],
 }
