+++ tsconfig-common.json
@@ -25,6 +25,7 @@
 
         /* Completeness */
         "skipDefaultLibCheck": true, // it's safe to assume that built-in .d.ts files are correct
-        "skipLibCheck": false
+        "skipLibCheck": true,
+        "typeRoots": ["./node_modules/@types", "**/types"],
     }
 }
