+++ tsconfig.json
@@ -2,6 +2,7 @@
 /* Visit https://aka.ms/tsconfig to read more about this file */
 {
     "exclude": ["**/dist"],
+    "include": ["packages/*/src", "build-scripts"],
     "extends": "./tsconfig-common.json",
     "compilerOptions": {
         "noEmit": true,
